% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp.R
\name{ibp}
\alias{ibp}
\title{Define an Indian Buffet Process (IBP) Distribution for Feature Allocations}
\usage{
ibp(mass, x)
}
\arguments{
\item{mass}{The mass (a.k.a., concentration) parameter.}

\item{x}{A character vector giving the labels of the items to place in a feature allocation, or an integer giving the number of items.}
}
\value{
An object representing an Indian Buffet Process (IBP) feature allocation distribution.
}
\description{
This function specifies an Indian Buffet Process (IBP), which is a
distribution over feature allocations.
}
\examples{
ibp(1,5)
ibp(1,c("CA","WI","NE","NY","UT"))

}
